requirements.txt
# PDF Processing with LlamaParse
llama-parse>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0

# Data Cleaning with OpenAI
openai>=1.0.0 